:truck: Separate LDAP trees
